pyproject.toml
[project]
name = "chicken"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "captcha>=0.7.1",
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "keras>=3.10.0",
    "matplotlib>=3.7.0",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "scikit-learn>=1.3.0",
    "seaborn>=0.13.2",
    "selenium>=4.34.0",
    "tensorflow>=2.19.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
